requirements.txt
praw
python-pushover
PyMySQL
